requirements.txt
flask
requests
jsonify
google-generativeai
serpapi
Openai
gemini-embeddings
